# packages/api-client/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/users/me:
    get:
      operationId: getMe
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUserDto'
      security:
        - bearer: []
      summary: Return the authenticated user profile
      tags:
        - users
    patch:
      operationId: updateMe
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileBodyDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUserDto'
      security:
        - bearer: []
      summary: Update the authenticated user profile
      tags:
        - users
...
  /api/v1/seat-layouts:
    get:
      operationId: listSeatLayouts
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SeatLayoutDto'
      summary: List all seat layouts (public)
      tags:
        - seat-layouts
    post:
      operationId: createSeatLayout
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSeatLayoutBodyDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatLayoutDto'
      security:
        - bearer: []
      summary: Create a seat layout with seats in one transaction (admin)
      tags:
        - seat-layouts
  /api/v1/seat-layouts/{id}:
    get:
      operationId: getSeatLayout
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatLayoutDto'
      summary: Get a seat layout with seats (public)
      tags:
        - seat-layouts
    patch:
      operationId: updateSeatLayout
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSeatLayoutBodyDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatLayoutDto'
      security:
        - bearer: []
      summary: Update seat layout metadata (admin)
      tags:
        - seat-layouts
    delete:
      operationId: deleteSeatLayout
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '204':
          description: ''
      security:
        - bearer: []
      summary: Delete a seat layout (admin) — 409 if vehicles reference it
      tags:
        - seat-layouts
...
  /api/v1/vehicles:
    get:
      operationId: listVehicles
      parameters:
        - name: companyId
          required: false
          in: query
          schema:
            maximum: 9007199254740991
            exclusiveMinimum: true
            type: integer
            minimum: 0
        - name: type
          required: false
          in: query
          schema:
            minLength: 1
            maxLength: 50
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleDto'
      summary: Search vehicles (public)
      tags:
        - vehicles
    post:
      operationId: createVehicle
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVehicleBodyDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleDto'
      security:
        - bearer: []
      summary: Create a vehicle (admin or operator-scoped)
      tags:
        - vehicles
  /api/v1/vehicles/{id}:
    get:
      operationId: getVehicle
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleDto'
      summary: Get a single vehicle (public)
      tags:
        - vehicles
    patch:
      operationId: updateVehicle
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVehicleBodyDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleDto'
      security:
        - bearer: []
      summary: Update a vehicle (admin or owning operator)
      tags:
        - vehicles
    delete:
      operationId: deleteVehicle
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '204':
          description: ''
      security:
        - bearer: []
      summary: Delete a vehicle (admin or owning operator) — 409 if active trips exist
      tags:
        - vehicles
info:
  title: Bookee API
  description: Multi-operator trip booking platform API
  version: 0.1.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    PublicUserDto:
      type: object
      properties:
        id:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
          pattern: ^(?!\.)(?!.*\.\.)([A-Za-z0-9_'+\-\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$
        role:
          type: string
          enum:
            - customer
            - operator
            - driver
            - admin
        operatorId:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - phone
        - email
        - role
        - operatorId
        - createdAt
        - updatedAt
    UpdateProfileBodyDto:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        phone:
          type: string
          minLength: 8
          maxLength: 20
          pattern: ^[+\d\s-]+$
        email:
          type: string
          maxLength: 254
          format: email
          pattern: ^(?!\.)(?!.*\.\.)([A-Za-z0-9_'+\-\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$
...
    SeatLayoutDto:
      type: object
      properties:
        id:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        name:
          type: string
        rows:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        cols:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        seats:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                minimum: -9007199254740991
                maximum: 9007199254740991
              layoutId:
                type: integer
                minimum: -9007199254740991
                maximum: 9007199254740991
              code:
                type: string
              floor:
                type: integer
                minimum: -9007199254740991
                maximum: 9007199254740991
              row:
                type: integer
                minimum: -9007199254740991
                maximum: 9007199254740991
              col:
                type: integer
                minimum: -9007199254740991
                maximum: 9007199254740991
            required:
              - id
              - layoutId
              - code
              - floor
              - row
              - col
      required:
        - id
        - name
        - rows
        - cols
    CreateSeatLayoutBodyDto:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        rows:
          type: integer
          exclusiveMinimum: true
          maximum: 9007199254740991
          minimum: 0
        cols:
          type: integer
          exclusiveMinimum: true
          maximum: 9007199254740991
          minimum: 0
        seats:
          minItems: 1
          type: array
          items:
            type: object
            properties:
              code:
                type: string
                minLength: 1
                maxLength: 20
              floor:
                default: 1
                type: integer
                minimum: 1
                maximum: 9007199254740991
              row:
                type: integer
                minimum: 1
                maximum: 9007199254740991
              col:
                type: integer
                minimum: 1
                maximum: 9007199254740991
            required:
              - code
              - row
              - col
      required:
        - name
        - rows
        - cols
        - seats
    UpdateSeatLayoutBodyDto:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        rows:
          type: integer
          exclusiveMinimum: true
          maximum: 9007199254740991
          minimum: 0
        cols:
          type: integer
          exclusiveMinimum: true
          maximum: 9007199254740991
          minimum: 0
...
    VehicleDto:
      type: object
      properties:
        id:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        companyId:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        plateNumber:
          type: string
        type:
          type: string
        seatLayoutId:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        totalSeats:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
      required:
        - id
        - companyId
        - plateNumber
        - type
        - seatLayoutId
        - totalSeats
    CreateVehicleBodyDto:
      type: object
      properties:
        companyId:
          type: integer
          exclusiveMinimum: true
          maximum: 9007199254740991
          minimum: 0
        plateNumber:
          type: string
          minLength: 1
          maxLength: 20
        type:
          type: string
          minLength: 1
          maxLength: 50
        seatLayoutId:
          type: integer
          exclusiveMinimum: true
          maximum: 9007199254740991
          minimum: 0
        totalSeats:
          type: integer
          exclusiveMinimum: true
          maximum: 9007199254740991
          minimum: 0
      required:
        - companyId
        - plateNumber
        - type
        - seatLayoutId
        - totalSeats
    UpdateVehicleBodyDto:
      type: object
      properties:
        plateNumber:
          type: string
          minLength: 1
          maxLength: 20
        type:
          type: string
          minLength: 1
          maxLength: 50
        seatLayoutId:
          type: integer
          exclusiveMinimum: true
          maximum: 9007199254740991
          minimum: 0
        totalSeats:
          type: integer
          exclusiveMinimum: true
          maximum: 9007199254740991
          minimum: 0